pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "smart-logger"
version = "1.0.0"
description = "AI-powered CLI tool to log work to Jira and Notion"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "google-genai>=1.0.0",
    "jira>=3.5.0",
    "playwright>=1.40.0",
    "python-dotenv>=1.0.0",
    "requests>=2.31.0",
    "rich>=13.0.0",
    "typer>=0.9.0",
]

[project.scripts]
smart-log = "main:cli"

[tool.setuptools]
py-modules = ["main", "notion_auth", "notion_form", "memory_bank", "ai_orchestrator", "genai_client"]
